Fix Bluebird warning
The problem here is that the `thenable` passed when
`this.dontProcessStalledJobs` is a resolved Promise. However, by the
specification of Promises, the `thenable` must be a _function_ that
_resolves to_ a Promise.

The fix is to pass `Promise.resolve` instead, without
invoking. `Promise.resolve` is an utility function that returns a resolved
Promise.

No state is leaked to subsequent thenable functions (this is by design
of Promises) so doing this is safe.

This patch is to be reverted once we migrate to 1.0rc and implement a
proper opt-in/opt-out for processStalledJobs feature.